refactor(health): pass repo & node.Netork into admin
