+++ api/admin/admin.go
@@ -8,21 +8,22 @@ package admin
 import (
 	"log/slog"
 	"net/http"
+	"time"
 
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
+	"github.com/vechain/thor/v2/api/admin/health"
 	"github.com/vechain/thor/v2/api/admin/loglevel"
-	"github.com/vechain/thor/v2/health"
-
-	healthAPI "github.com/vechain/thor/v2/api/admin/health"
+	"github.com/vechain/thor/v2/api/node"
+	"github.com/vechain/thor/v2/chain"
 )
 
-func New(logLevel *slog.LevelVar, health *health.Health) http.HandlerFunc {
+func New(logLevel *slog.LevelVar, repo *chain.Repository, node node.Network, blockInterval time.Duration) http.HandlerFunc {
 	router := mux.NewRouter()
 	subRouter := router.PathPrefix("/admin").Subrouter()
 
 	loglevel.New(logLevel).Mount(subRouter, "/loglevel")
-	healthAPI.New(health).Mount(subRouter, "/health")
+	health.New(repo, node, blockInterval).Mount(subRouter, "/health")
 
 	handler := handlers.CompressHandler(router)
