+++ cmd/thor/solo/solo.go
@@ -22,7 +22,6 @@ import (
 	"github.com/vechain/thor/v2/cmd/thor/bandwidth"
 	"github.com/vechain/thor/v2/co"
 	"github.com/vechain/thor/v2/genesis"
-	"github.com/vechain/thor/v2/health"
 	"github.com/vechain/thor/v2/log"
 	"github.com/vechain/thor/v2/logdb"
 	"github.com/vechain/thor/v2/packer"
@@ -44,7 +43,6 @@ type Solo struct {
 	txPool        *txpool.TxPool
 	packer        *packer.Packer
 	logDB         *logdb.LogDB
-	health        *health.Health
 	gasLimit      uint64
 	bandwidth     bandwidth.Bandwidth
 	blockInterval uint64
@@ -57,7 +55,6 @@ func New(
 	repo *chain.Repository,
 	stater *state.Stater,
 	logDB *logdb.LogDB,
-	health *health.Health,
 	txPool *txpool.TxPool,
 	gasLimit uint64,
 	onDemand bool,
@@ -76,7 +73,6 @@ func New(
 			&genesis.DevAccounts()[0].Address,
 			forkConfig),
 		logDB:         logDB,
-		health:        health,
 		gasLimit:      gasLimit,
 		blockInterval: blockInterval,
 		skipLogs:      skipLogs,
@@ -213,8 +209,6 @@ func (s *Solo) packing(pendingTxs tx.Transactions, onDemand bool) error {
 		"id", fmt.Sprintf("[#%v…%x]", block.Number(blockID), blockID[28:]),
 	)
 	logger.Debug(b.String())
-
-	s.health.NewBestBlock(b.Header().ID())
 
 	return nil
 }
