+++ comm/communicator.go
@@ -20,7 +20,6 @@ import (
 	"github.com/vechain/thor/v2/chain"
 	"github.com/vechain/thor/v2/co"
 	"github.com/vechain/thor/v2/comm/proto"
-	"github.com/vechain/thor/v2/health"
 	"github.com/vechain/thor/v2/log"
 	"github.com/vechain/thor/v2/thor"
 	"github.com/vechain/thor/v2/tx"
@@ -40,13 +39,12 @@ type Communicator struct {
 	newBlockFeed   event.Feed
 	announcementCh chan *announcement
 	feedScope      event.SubscriptionScope
-	health         *health.Health
 	goes           co.Goes
 	onceSynced     sync.Once
 }
 
 // New create a new Communicator instance.
-func New(repo *chain.Repository, txPool *txpool.TxPool, health *health.Health) *Communicator {
+func New(repo *chain.Repository, txPool *txpool.TxPool) *Communicator {
 	ctx, cancel := context.WithCancel(context.Background())
 	return &Communicator{
 		repo:           repo,
@@ -54,7 +52,6 @@ func New(repo *chain.Repository, txPool *txpool.TxPool, health *health.Health) *
 		ctx:            ctx,
 		cancel:         cancel,
 		peerSet:        newPeerSet(),
-		health:         health,
 		syncedCh:       make(chan struct{}),
 		announcementCh: make(chan *announcement),
 	}
@@ -121,8 +118,6 @@ func (c *Communicator) Sync(ctx context.Context, handler HandleBlockStream) {
 			if isSynced() {
 				delay = syncInterval
 				c.onceSynced.Do(func() {
-					// once off - after a bootstrap the syncedCh trigger the peers.syncTxs
-					c.health.BootstrapStatus(true)
 					close(c.syncedCh)
 				})
 			}
