+++ thorclient/httpclient/http_request.go
@@ -21,7 +21,7 @@ func (c *Client) httpRequest(method, url string, payload io.Reader) ([]byte, err
 		return nil, err
 	}
 	if !statusCodeIs2xx(statusCode) {
-		return nil, fmt.Errorf("http error - Status Code %d - %s - %w", statusCode, body, common.ErrNot200Status)
+		return nil, fmt.Errorf("http error - status Code %d - %s - %w", statusCode, body, common.ErrNot200Status)
 	}
 	return body, nil
 }
